fix: merge로 인한 테스트코드 오류 해결
